jingle-media-rtp: Use guint to make compiler happier with GUINT_TO_POINTER

+++ src/jingle-media-rtp.h
@@ -61,7 +61,7 @@ struct _GabbleJingleMediaRtp {
 };
 
 typedef struct {
-  guint8 id;
+  guint id;
   gchar *name;
   guint clockrate;
   guint channels;
